+++ database/upgrade/upgrade4.8.0RC.sql
@@ -1,0 +1,15 @@
+-- ---------------------------------------------------------------------------------------------------------------------
+-- WARNING: BACKUP YOUR DATABASE BEFORE UPGRADING
+-- THIS SCRIPT IS ONLY FOR UPGRADING 4.7.0 TO 4.8.0
+-- THE CURRENT VERSION CAN BE FOUND AT `myems_system_db`.`tbl_versions`
+-- ---------------------------------------------------------------------------------------------------------------------
+
+START TRANSACTION;
+
+ALTER TABLE myems_system_db.tbl_tariffs_timeofuses MODIFY COLUMN price decimal(18,5) NOT NULL;
+
+-- UPDATE VERSION NUMBER
+UPDATE `myems_system_db`.`tbl_versions` SET version='4.8.0RC', release_date='2024-08-08' WHERE id=1;
+
+COMMIT;
+