+++ CHANGELOG.md
@@ -14,10 +14,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - added microgrid renevue report
 - added microgrid carbon report
 - added period type list automaticly adjustment logic in myems-web
-- 
 ### Changed
-- updated tariff price to keep 5 decimal places in myems-admin
-- 
+- updated tariff price to keep 5 decimal places in database and myems-admin
 ### Fixed
 - aligned timestamps of parameters for microgrid details report
 - aligned timestamps of parameters for energy storage power station details report